+++ test/test_main_config_files.py
@@ -5,20 +5,11 @@ import os
 import tempfile
 from test.markdown_scanner import MarkdownScanner
 
+import yaml
+
 from .utils import write_temporary_configuration
 
-
-def test_markdown_with_config_no_config():
-    """
-    Test to make sure that
-
-    This function shadows
-    test_api_scan_string_test_bad_file_due_to_no_disables
-    """
-
-    # Arrange
-    scanner = MarkdownScanner()
-    xxx = """# This is a document
+__TEST_DOCUMENT = """# This is a document
 
 * a list
   - a sublist
@@ -27,6 +18,18 @@ def test_markdown_with_config_no_config():
 this is a very long line
 """
 
+
+def test_markdown_with_config_no_config():
+    """
+    Test to make sure that we have a baseline for the other configuration tests.
+
+    This function shadows
+    test_api_scan_string_test_bad_file_due_to_no_disables
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
     supplied_arguments = [
         "--strict-config",
         "scan-stdin",
@@ -38,7 +41,7 @@ this is a very long line
 
     # Act
     execute_results = scanner.invoke_main(
-        arguments=supplied_arguments, standard_input_to_use=xxx
+        arguments=supplied_arguments, standard_input_to_use=stdin_to_use
     )
 
     # Assert
@@ -49,7 +52,8 @@ this is a very long line
 
 def test_markdown_with_config_general_command_line():
     """
-    Test to make sure that
+    Test to make sure that we can disable a rule from the command line without
+    any configuration file.
 
     This test shadows
     test_api_scan_string_test_good_file_after_disables
@@ -57,15 +61,7 @@ def test_markdown_with_config_general_command_line():
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
-
+    stdin_to_use = __TEST_DOCUMENT
     supplied_arguments = [
         "--strict-config",
         "-d",
@@ -79,7 +75,7 @@ this is a very long line
 
     # Act
     execute_results = scanner.invoke_main(
-        arguments=supplied_arguments, standard_input_to_use=xxx
+        arguments=supplied_arguments, standard_input_to_use=stdin_to_use
     )
 
     # Assert
@@ -90,20 +86,14 @@ this is a very long line
 
 def test_markdown_with_config_general_command_line_and_specific_command_line():
     """
-    Test to make sure that
+    Test to make sure that we can disable a rule specifically and enable a
+    rule through a configuration setting, with the disable winning as it is
+    the more specific configuration.
     """
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
-
+    stdin_to_use = __TEST_DOCUMENT
     supplied_arguments = [
         "--strict-config",
         "-d",
@@ -119,8 +109,304 @@ this is a very long line
 
     # Act
     execute_results = scanner.invoke_main(
-        arguments=supplied_arguments, standard_input_to_use=xxx
+        arguments=supplied_arguments, standard_input_to_use=stdin_to_use
     )
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_json_configuration_file():
+    """
+    Test to make sure that we can supply configuration via a JSON configuration file.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": False}},
+    }
+
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        configuration_file = None
+        try:
+            configuration_file = write_temporary_configuration(
+                specified_configuration, file_name="myconfig", directory=tmp_dir_path
+            )
+            supplied_arguments = [
+                "-c",
+                configuration_file,
+                "scan-stdin",
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(
+                arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+            )
+        finally:
+            if configuration_file and os.path.exists(configuration_file):
+                os.remove(configuration_file)
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_yaml_configuration_file():
+    """
+    Test to make sure that we can supply configuration via a YAML configuration file.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": False}},
+    }
+
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        configuration_file = None
+        try:
+            configuration_file = write_temporary_configuration(
+                yaml.dump(specified_configuration),
+                file_name="myconfig",
+                directory=tmp_dir_path,
+            )
+            supplied_arguments = [
+                "--stack-trace",
+                "-c",
+                configuration_file,
+                "scan-stdin",
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            execute_results = scanner.invoke_main(
+                arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+            )
+        finally:
+            if configuration_file and os.path.exists(configuration_file):
+                os.remove(configuration_file)
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_default_yaml():
+    """
+    Test to make sure that we can supply configuration via a YAML configuration file.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_yaml_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": False}},
+    }
+
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        default_yaml_configuration_file = None
+        try:
+            default_yaml_configuration_file = write_temporary_configuration(
+                yaml.dump(specified_yaml_configuration),
+                file_name=".pymarkdown.yaml",
+                directory=tmp_dir_path,
+            )
+            supplied_arguments = [
+                "scan-stdin",
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            old_current_working_directory = os.getcwd()
+            try:
+                os.chdir(tmp_dir_path)
+                execute_results = scanner.invoke_main(
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+                )
+            finally:
+                os.chdir(old_current_working_directory)
+        finally:
+            if default_yaml_configuration_file and os.path.exists(
+                default_yaml_configuration_file
+            ):
+                os.remove(default_yaml_configuration_file)
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_default_yml():
+    """
+    Test to make sure that we can supply configuration via a YML configuration file.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_yaml_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": False}},
+    }
+
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        default_yaml_configuration_file = None
+        try:
+            default_yaml_configuration_file = write_temporary_configuration(
+                yaml.dump(specified_yaml_configuration),
+                file_name=".pymarkdown.yml",
+                directory=tmp_dir_path,
+            )
+            supplied_arguments = [
+                "scan-stdin",
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            old_current_working_directory = os.getcwd()
+            try:
+                os.chdir(tmp_dir_path)
+                execute_results = scanner.invoke_main(
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+                )
+            finally:
+                os.chdir(old_current_working_directory)
+        finally:
+            if default_yaml_configuration_file and os.path.exists(
+                default_yaml_configuration_file
+            ):
+                os.remove(default_yaml_configuration_file)
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_default_json_beats_default_yaml():
+    """
+    Test to make sure that we can supply configuration via a YAML configuration file.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_json_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": False}},
+    }
+    specified_yaml_configuration = {
+        "mode": {"strict-config": True},
+        "plugins": {"md004": {"enabled": True}},
+    }
+
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        default_json_configuration_file = None
+        default_yaml_configuration_file = None
+        try:
+            default_json_configuration_file = write_temporary_configuration(
+                specified_json_configuration,
+                file_name=".pymarkdown",
+                directory=tmp_dir_path,
+            )
+            default_yaml_configuration_file = write_temporary_configuration(
+                yaml.dump(specified_yaml_configuration),
+                file_name=".pymarkdown.yaml",
+                directory=tmp_dir_path,
+            )
+            supplied_arguments = [
+                "scan-stdin",
+            ]
+
+            expected_return_code = 0
+            expected_output = ""
+            expected_error = ""
+
+            # Act
+            old_current_working_directory = os.getcwd()
+            try:
+                os.chdir(tmp_dir_path)
+                execute_results = scanner.invoke_main(
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+                )
+            finally:
+                os.chdir(old_current_working_directory)
+        finally:
+            if default_yaml_configuration_file and os.path.exists(
+                default_yaml_configuration_file
+            ):
+                os.remove(default_yaml_configuration_file)
+            if default_json_configuration_file and os.path.exists(
+                default_json_configuration_file
+            ):
+                os.remove(default_json_configuration_file)
+
+    # Assert
+    execute_results.assert_results(
+        expected_output, expected_error, expected_return_code
+    )
+
+
+def test_markdown_with_config_bad_json_and_yaml_configuration_file():
+    """
+    Test to make sure that we error out properly with a file that is
+    not JSON and is not YAML.
+    """
+
+    # Arrange
+    scanner = MarkdownScanner()
+    stdin_to_use = __TEST_DOCUMENT
+    specified_configuration = """hallo: 1
+bye
+"""
+    with tempfile.TemporaryDirectory() as tmp_dir_path:
+        configuration_file = None
+        try:
+            configuration_file = write_temporary_configuration(
+                specified_configuration, file_name="myconfig", directory=tmp_dir_path
+            )
+            supplied_arguments = [
+                "--stack-trace",
+                "-c",
+                configuration_file,
+                "scan-stdin",
+            ]
+
+            expected_return_code = 1
+            expected_output = ""
+            expected_error = f"Specified configuration file '{configuration_file}' was not parseable as a JSON file or a YAML file."
+
+            # Act
+            execute_results = scanner.invoke_main(
+                arguments=supplied_arguments, standard_input_to_use=stdin_to_use
+            )
+        finally:
+            if configuration_file and os.path.exists(configuration_file):
+                os.remove(configuration_file)
 
     # Assert
     execute_results.assert_results(
@@ -130,22 +416,17 @@ this is a very long line
 
 def test_markdown_with_config_specific_command_line_and_configuration_file():
     """
-    Test to make sure that
+    Test to make sure that if a command line setting and a configuration
+    file setting are applied, that the command line setting wins as it is
+    more specific.
     """
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
+    stdin_to_use = __TEST_DOCUMENT
     specified_configuration = {
         "mode": {"strict-config": True},
-        "plugins": {"md004": {"enabled": False}},
+        "plugins": {"md004": {"enabled": True}},
     }
 
     with tempfile.TemporaryDirectory() as tmp_dir_path:
@@ -169,7 +450,7 @@ this is a very long line
 
             # Act
             execute_results = scanner.invoke_main(
-                arguments=supplied_arguments, standard_input_to_use=xxx
+                arguments=supplied_arguments, standard_input_to_use=stdin_to_use
             )
         finally:
             if configuration_file and os.path.exists(configuration_file):
@@ -183,19 +464,13 @@ this is a very long line
 
 def test_markdown_with_config_configuration_file_and_default_configuration_file():
     """
-    Test to make sure that
+    Test to make sure that any default configuration is overridden by an explicitly
+    named configuration file.
     """
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
+    stdin_to_use = __TEST_DOCUMENT
     specified_configuration = {
         "mode": {"strict-config": True},
         "plugins": {"md004": {"enabled": False}},
@@ -231,7 +506,7 @@ this is a very long line
             try:
                 os.chdir(tmp_dir_path)
                 execute_results = scanner.invoke_main(
-                    arguments=supplied_arguments, standard_input_to_use=xxx
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
                 )
             finally:
                 os.chdir(old_current_working_directory)
@@ -251,19 +526,13 @@ this is a very long line
 
 def test_markdown_with_default_configuration_file_and_alternate_configuration_file():
     """
-    Test to make sure that
+    Test to make sure that a default configuration file takes precedence over a
+    Python standard configuration file.
     """
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
+    stdin_to_use = __TEST_DOCUMENT
     default_configuration = {
         "mode": {"strict-config": True},
         "plugins": {"md004": {"enabled": False}},
@@ -299,7 +568,7 @@ plugins.md004.enabled = true
             try:
                 os.chdir(tmp_dir_path)
                 execute_results = scanner.invoke_main(
-                    arguments=supplied_arguments, standard_input_to_use=xxx
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
                 )
             finally:
                 os.chdir(old_current_working_directory)
@@ -321,19 +590,13 @@ plugins.md004.enabled = true
 
 def test_markdown_with_alternate_configuration_file():
     """
-    Test to make sure that
+    Test to make sure that we can load configuration from a Python standard
+    configuration file.
     """
 
     # Arrange
     scanner = MarkdownScanner()
-    xxx = """# This is a document
-
-* a list
-  - a sublist
-  - a very long sublist item
-
-this is a very long line
-"""
+    stdin_to_use = __TEST_DOCUMENT
     alternate_configuration = """
 [tool.pymarkdown]
 plugins.md004.enabled = false
@@ -361,7 +624,7 @@ plugins.md004.enabled = false
             try:
                 os.chdir(tmp_dir_path)
                 execute_results = scanner.invoke_main(
-                    arguments=supplied_arguments, standard_input_to_use=xxx
+                    arguments=supplied_arguments, standard_input_to_use=stdin_to_use
                 )
             finally:
                 os.chdir(old_current_working_directory)
@@ -377,9 +640,10 @@ plugins.md004.enabled = false
     )
 
 
-def test_markdown_with_pyproject_configuration_file_xxx():
+def test_markdown_with_pyproject_configuration_file_with_bad_log_file_value_type():
     """
-    Test to make sure that
+    Test to make sure that a standard python configuration file can be used and
+    that an error with value types is reported.
     """
 
     # Arrange